"ads2_m8_a1_1: 6/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-2 Assignments/Module 8/m8/1. Shortest Paths/Solution.java
@@ -24,7 +24,7 @@ final class Solution {
 		for (int k = 0; k < querynum; k++) {
 			String[] querydetails = scan.nextLine().split(" ");
 			DijkstraSP dijkobj = new DijkstraSP(edgegraph, stationlist.indexOf(querydetails[0]));
-			System.out.println(dijkobj.distTo(stationlist.indexOf(querydetails[1])));
+			System.out.println((long)dijkobj.distTo(stationlist.indexOf(querydetails[1])));
 		}
 	}
 }
